+++ tests/data/finder_charts/finder_chart.pdf
